alzheimer-diagnosis-pred-service/requirements.txt
fastapi==0.88.0
pillow==9.3.0
uvicorn==0.20.0
python-multipart==0.0.5
keras==2.11.0
opencv-python==4.7.0.68
numpy==1.24.1
pydantic==1.10.4
tensorflow==2.11